move RPC server implementation to server crate
