+++ proto/src/examples/dummy_client.rs
@@ -1,9 +1,11 @@
+#![allow(dead_code, unused)]
+
+use prost_types::Timestamp;
 use std::net::SocketAddr;
+use tonic::transport::Channel;
 
 use crate::ollyllm::ollyllm_service_client::OllyllmServiceClient;
 use crate::ollyllm::{ReportSpanRequest, Span, TestExecutionRequest, VersionedTest};
-use prost_types::Timestamp;
-use tonic::transport::Channel;
 
 pub struct Client {
     client: OllyllmServiceClient<Channel>,
@@ -52,68 +54,5 @@ impl Client {
                 test_input: Vec::new(),
             });
         self.client.queue_test(test_request).await
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::server::RpcServer;
-    use std::net::{SocketAddr, TcpListener};
-    use tokio::{sync::oneshot, task::JoinHandle};
-
-    struct Fresh;
-
-    impl Fresh {
-        fn socker_addr() -> SocketAddr {
-            let listener: TcpListener = TcpListener::bind("127.0.0.1:0").unwrap();
-            listener.local_addr().unwrap()
-        }
-    }
-
-    #[tokio::test]
-    async fn test_queue_span_creation_request() {
-        let addr: SocketAddr = Fresh::socker_addr();
-        let (tx, rx): (oneshot::Sender<()>, oneshot::Receiver<()>) = oneshot::channel();
-
-        let server_handle: JoinHandle<()> = tokio::spawn(async move {
-            let server: RpcServer = RpcServer::new(addr.clone()).await;
-
-            tx.send(()).unwrap();
-            server.serve().await.unwrap();
-        });
-
-        rx.await.unwrap();
-
-        let mut client: Client = Client::new(addr).await.unwrap();
-        let response: Result<tonic::Response<()>, tonic::Status> =
-            client.send_dummy_span_creation_request().await;
-
-        assert!(response.is_ok());
-
-        server_handle.abort();
-    }
-
-    #[tokio::test]
-    async fn test_queue_test_execution_request() {
-        let addr: SocketAddr = Fresh::socker_addr();
-        let (tx, rx): (oneshot::Sender<()>, oneshot::Receiver<()>) = oneshot::channel();
-
-        let server_handle: JoinHandle<()> = tokio::spawn(async move {
-            let server: RpcServer = RpcServer::new(addr.clone()).await;
-
-            tx.send(()).unwrap();
-            server.serve().await.unwrap();
-        });
-
-        rx.await.unwrap();
-
-        let mut client: Client = Client::new(addr).await.unwrap();
-        let response: Result<tonic::Response<()>, tonic::Status> =
-            client.send_dummy_test_execution_request().await;
-
-        assert!(response.is_ok());
-
-        server_handle.abort();
     }
 }
