+++ proto/src/lib.rs
@@ -2,4 +2,3 @@ pub mod ollyllm {
     include!("../gen/src/ollyllm.v1.rs");
 }
 pub mod examples;
-pub mod server;
